+++ pkg/api/contacts.go
@@ -2,47 +2,18 @@ package api
 
 import (
 	"database/sql"
-	"time"
 
 	"github.com/go-fuego/fuego"
 	"github.com/nzoschke/codon/pkg/db"
-	"github.com/nzoschke/codon/pkg/sql/models"
+	"github.com/nzoschke/codon/pkg/models"
 	"github.com/nzoschke/codon/pkg/sql/q"
 	"github.com/olekukonko/errors"
 )
 
-type Contact struct {
-	CreatedAt time.Time `json:"created_at"`
-	Email     string    `json:"email"`
-	ID        int       `json:"id"`
-	Info      Info      `json:"info"`
-	Name      string    `json:"name"`
-	Phone     string    `json:"phone"`
-	UpdatedAt time.Time `json:"updated_at"`
-}
-
-type ContactCreateIn struct {
-	Email string `form:"email" json:"email"`
-	Info  Info   `form:"info" json:"info"`
-	Name  string `form:"name" json:"name"`
-	Phone string `form:"phone" json:"phone"`
-}
-
-type ContactUpdateIn struct {
-	Email string `json:"email"`
-	Info  Info   `json:"info"`
-	Name  string `json:"name"`
-	Phone string `json:"phone"`
-}
-
-type Info struct {
-	Age int `json:"age"`
-}
-
 func Contacts(s *fuego.Server, db db.DB) {
 	g := fuego.Group(s, "/contacts")
 
-	fuego.Get(g, "", func(c fuego.ContextNoBody) ([]Contact, error) {
+	fuego.Get(g, "", func(c fuego.ContextNoBody) ([]models.Contact, error) {
 		ctx := c.Context()
 
 		conn, put, err := db.Take(ctx)
@@ -56,13 +27,13 @@ func Contacts(s *fuego.Server, db db.DB) {
 			return nil, errors.WithStack(err)
 		}
 
-		out := []Contact{}
+		out := []models.Contact{}
 		for _, r := range rows {
-			out = append(out, Contact{
+			out = append(out, models.Contact{
 				CreatedAt: r.CreatedAt,
 				Email:     r.Email,
 				ID:        int(r.Id),
-				Info:      Info(r.Info),
+				Info:      models.Info(r.Info),
 				Name:      r.Name,
 				Phone:     r.Phone,
 				UpdatedAt: r.UpdatedAt,
@@ -75,17 +46,17 @@ func Contacts(s *fuego.Server, db db.DB) {
 		fuego.OptionSummary("list"),
 	)
 
-	fuego.Post(g, "", func(c fuego.ContextWithBody[ContactCreateIn]) (Contact, error) {
+	fuego.Post(g, "", func(c fuego.ContextWithBody[models.ContactCreateIn]) (models.Contact, error) {
 		ctx := c.Context()
 
 		in, err := c.Body()
 		if err != nil {
-			return Contact{}, err
+			return models.Contact{}, err
 		}
 
 		conn, put, err := db.Take(ctx)
 		if err != nil {
-			return Contact{}, errors.WithStack(err)
+			return models.Contact{}, errors.WithStack(err)
 		}
 		defer put()
 
@@ -96,14 +67,14 @@ func Contacts(s *fuego.Server, db db.DB) {
 			Phone: in.Phone,
 		})
 		if err != nil {
-			return Contact{}, errors.WithStack(err)
+			return models.Contact{}, errors.WithStack(err)
 		}
 
-		out := Contact{
+		out := models.Contact{
 			CreatedAt: r.CreatedAt,
 			Email:     r.Email,
 			ID:        int(r.Id),
-			Info:      Info(r.Info),
+			Info:      r.Info,
 			Name:      r.Name,
 			Phone:     r.Phone,
 			UpdatedAt: r.UpdatedAt,
@@ -137,28 +108,28 @@ func Contacts(s *fuego.Server, db db.DB) {
 		fuego.OptionSummary("delete"),
 	)
 
-	fuego.Get(g, "/{id}", func(c fuego.ContextNoBody) (Contact, error) {
+	fuego.Get(g, "/{id}", func(c fuego.ContextNoBody) (models.Contact, error) {
 		ctx := c.Context()
 
 		conn, put, err := db.Take(ctx)
 		if err != nil {
-			return Contact{}, errors.WithStack(err)
+			return models.Contact{}, errors.WithStack(err)
 		}
 		defer put()
 
 		r, err := q.ContactRead(conn, int64(c.PathParamInt("id")))
 		if err != nil {
 			if err == sql.ErrNoRows {
-				return Contact{}, fuego.NotFoundError{}
+				return models.Contact{}, fuego.NotFoundError{}
 			}
-			return Contact{}, errors.WithStack(err)
+			return models.Contact{}, errors.WithStack(err)
 		}
 
-		out := Contact{
+		out := models.Contact{
 			CreatedAt: r.CreatedAt,
 			Email:     r.Email,
 			ID:        int(r.Id),
-			Info:      Info(r.Info),
+			Info:      r.Info,
 			Name:      r.Name,
 			Phone:     r.Phone,
 			UpdatedAt: r.UpdatedAt,
@@ -170,19 +141,19 @@ func Contacts(s *fuego.Server, db db.DB) {
 		fuego.OptionSummary("get"),
 	)
 
-	fuego.Put(g, "/{id}", func(c fuego.ContextWithBody[ContactUpdateIn]) (Contact, error) {
+	fuego.Put(g, "/{id}", func(c fuego.ContextWithBody[models.ContactUpdateIn]) (models.Contact, error) {
 		ctx := c.Request().Context()
 
 		id := int64(c.PathParamInt("id"))
 
 		in, err := c.Body()
 		if err != nil {
-			return Contact{}, err
+			return models.Contact{}, err
 		}
 
 		conn, put, err := db.Take(ctx)
 		if err != nil {
-			return Contact{}, errors.WithStack(err)
+			return models.Contact{}, errors.WithStack(err)
 		}
 		defer put()
 
@@ -194,22 +165,22 @@ func Contacts(s *fuego.Server, db db.DB) {
 			Phone: in.Phone,
 		})
 		if err != nil {
-			return Contact{}, errors.WithStack(err)
+			return models.Contact{}, errors.WithStack(err)
 		}
 
 		r, err := q.ContactRead(conn, id)
 		if err != nil {
 			if err == sql.ErrNoRows {
-				return Contact{}, fuego.NotFoundError{}
+				return models.Contact{}, fuego.NotFoundError{}
 			}
-			return Contact{}, errors.WithStack(err)
+			return models.Contact{}, errors.WithStack(err)
 		}
 
-		out := Contact{
+		out := models.Contact{
 			CreatedAt: r.CreatedAt,
 			Email:     r.Email,
 			ID:        int(r.Id),
-			Info:      Info(r.Info),
+			Info:      r.Info,
 			Name:      r.Name,
 			Phone:     r.Phone,
 			UpdatedAt: r.UpdatedAt,
