--- pkg/sql/models/contact.go
@@ -1,5 +1,0 @@
-package models
-
-type Info struct {
-	Age int `json:"age"`
-}
