+++ src/schema.d.ts
@@ -193,7 +193,7 @@ export interface operations {
       path?: never;
       cookie?: never;
     };
-    /** @description Request body for api.ContactCreateIn */
+    /** @description Request body for models.ContactCreateIn */
     requestBody: {
       content: {
         "*/*": components["schemas"]["ContactCreateIn"];
@@ -300,7 +300,7 @@ export interface operations {
       };
       cookie?: never;
     };
-    /** @description Request body for api.ContactUpdateIn */
+    /** @description Request body for models.ContactUpdateIn */
     requestBody: {
       content: {
         "*/*": components["schemas"]["ContactUpdateIn"];
